Convert pagerduty alert action config to json
