--- alerts/actions/pagerDutyTriggerEvent.conf
@@ -1,5 +1,0 @@
-[options]
-serviceKey = <yourapiservicekey>
-docs = {"alertcategory1": "https://website.with.runbooks.com/alert-runbook"}
-keywords = keyword1 keyword2
-clienturl = https://a.link.to.your.siem.com/alert
